move install logic into esmpkg
