+++ esmpkg/src/rollup-plugins/generateProcessPolyfill.ts
