+++ esmpkg/src/rollup-plugins/rollup-plugin-catch-fetch.ts
