+++ esmpkg/src/rollup-plugins/rollup-plugin-catch-unresolved.ts
