+++ esmpkg/src/rollup-plugins/rollup-plugin-strip-source-mapping.ts
