+++ esmpkg/src/types.ts
@@ -1,0 +1,49 @@
+export type DeepPartial<T> = {
+  [P in keyof T]?: T[P] extends Array<infer U>
+    ? Array<DeepPartial<U>>
+    : T[P] extends ReadonlyArray<infer U>
+    ? ReadonlyArray<DeepPartial<U>>
+    : DeepPartial<T[P]>;
+};
+
+export type EnvVarReplacements = Record<string, string | number | true>;
+
+export interface ImportMap {
+  imports: {[packageName: string]: string};
+}
+
+/**
+ * An install target represents information about a dependency to install.
+ * The specifier is the key pointing to the dependency, either as a package
+ * name or as an actual file path within node_modules. All other properties
+ * are metadata about what is actually being imported.
+ */
+export type InstallTarget = {
+  specifier: string;
+  all: boolean;
+  default: boolean;
+  namespace: boolean;
+  named: string[];
+};
+
+export type DependencyStats = {
+  size: number;
+  gzip: number;
+  brotli?: number;
+  delta?: number;
+};
+
+export type DependencyType = 'direct' | 'common';
+
+export type DependencyStatsMap = {
+  [filePath: string]: DependencyStats;
+};
+
+export type DependencyStatsOutput = Record<DependencyType, DependencyStatsMap>;
+
+export type LoggerLevel = 'debug' | 'info' | 'warn' | 'error' | 'silent'; // same as Pino
+export type LoggerEvent = 'debug' | 'info' | 'warn' | 'error';
+export interface LoggerOptions {
+  /** (optional) change name at beginning of line */
+  name?: string;
+}
