+++ snowpack/package.json
@@ -52,14 +52,9 @@
   },
   "dependencies": {
     "@babel/plugin-syntax-import-meta": "^7.10.4",
-    "@rollup/plugin-alias": "^3.0.1",
-    "@rollup/plugin-commonjs": "^15.0.0",
-    "@rollup/plugin-inject": "^4.0.2",
-    "@rollup/plugin-json": "^4.0.0",
-    "@rollup/plugin-node-resolve": "^9.0.0",
-    "@rollup/plugin-replace": "^2.3.3",
     "@snowpack/plugin-build-script": "^2.0.7",
     "@snowpack/plugin-run-script": "^2.1.2",
+    "esmpkg": "^0.0.1",
     "cacache": "^15.0.0",
     "cachedir": "^2.3.0",
     "chokidar": "^3.4.0",
@@ -77,7 +72,6 @@
     "glob": "^7.1.4",
     "got": "^11.1.4",
     "http-proxy": "^1.18.1",
-    "is-builtin-module": "^3.0.0",
     "isbinaryfile": "^4.0.6",
     "jsonschema": "^1.2.5",
     "kleur": "^4.1.1",
@@ -88,7 +82,6 @@
     "p-queue": "^6.6.1",
     "resolve-from": "^5.0.0",
     "rimraf": "^3.0.0",
-    "rollup": "^2.23.0",
     "rollup-plugin-node-polyfills": "^0.2.1",
     "signal-exit": "^3.0.3",
     "strip-comments": "^2.0.1",
