--- test/integration/source-pika-lockfile/package.json
@@ -1,20 +1,0 @@
-{
-  "private": true,
-  "version": "1.0.1",
-  "name": "@snowpack/test-source-pika-lockfile",
-  "description": "Test that the lockfile import URLs & version info are used",
-  "scripts": {
-    "testinstall": "snowpack"
-  },
-  "webDependencies": {
-    "async": "3.1.0"
-  },
-  "snowpack": {
-    "buildOptions": {
-      "minify": false
-    }
-  },
-  "devDependencies": {
-    "snowpack": "^2.7.0"
-  }
-}
