--- test/integration/source-pika-lockfile/snowpack.lock.json
@@ -1,5 +1,0 @@
-{
-  "imports": {
-    "async": "https://cdn.pika.dev/pin/async@v3.2.0-e41RUHPcKIMtuL3cP7sk/async.js"
-  }
-}
