--- test/integration/source-pika-types/package.json
@@ -1,20 +1,0 @@
-{
-  "private": true,
-  "version": "1.0.1",
-  "name": "@snowpack/test-source-pika-types",
-  "description": "Test that --install-types works as expected.",
-  "scripts": {
-    "testinstall": "snowpack --install-types"
-  },
-  "webDependencies": {
-    "async": "3.1.0"
-  },
-  "snowpack": {
-    "buildOptions": {
-      "minify": false
-    }
-  },
-  "devDependencies": {
-    "snowpack": "^2.7.0"
-  }
-}
