--- test/integration/source-pika-webdependencies/expected-lock.json
@@ -1,7 +1,0 @@
-{
-  "imports": {
-    "async": "https://cdn.pika.dev/pin/async@v3.0.1-o9C1cZwTqkk4zNrQqO6O/async.js",
-    "react": "https://cdn.pika.dev/pin/react@v16.13.1-ByypZEPVPs6cpkpGdpQK/react.js",
-    "react-dom": "https://cdn.pika.dev/pin/react-dom@v16.13.1-qv1YB4ZAVXJG84jEgino/react-dom.js"
-  }
-}
