--- test/integration/source-pika-webdependencies/package.json
@@ -1,22 +1,0 @@
-{
-  "private": true,
-  "version": "1.0.1",
-  "name": "@snowpack/test-source-pika-webdependencies",
-  "description": "Test that the lockfile import URLs & version info are used",
-  "scripts": {
-    "testinstall": "snowpack"
-  },
-  "webDependencies": {
-    "async": "~3.0.0",
-    "react": "16.13.1",
-    "react-dom": "16.13.1"
-  },
-  "snowpack": {
-    "buildOptions": {
-      "minify": false
-    }
-  },
-  "devDependencies": {
-    "snowpack": "^2.7.0"
-  }
-}
