--- test/integration/source-pika/expected-lock.json
@@ -1,5 +1,0 @@
-{
-  "imports": {
-    "async": "https://cdn.pika.dev/pin/async@v3.1.0-t2zPE8phAyA9kLXoTEKm/async.js"
-  }
-}
