--- test/integration/source-pika/package.json
@@ -1,20 +1,0 @@
-{
-  "private": true,
-  "version": "1.0.1",
-  "name": "@snowpack/test-source-pika",
-  "description": "Test that webDependencies work without a node_modules directory",
-  "scripts": {
-    "testinstall": "snowpack"
-  },
-  "webDependencies": {
-    "async": "3.1.0"
-  },
-  "snowpack": {
-    "buildOptions": {
-      "minify": false
-    }
-  },
-  "devDependencies": {
-    "snowpack": "^2.7.0"
-  }
-}
